requirements.txt
streamlit~=1.45.0
pillow~=11.2.1
PyPDF2~=3.0.1
numpy~=1.26.4
easyocr~=1.7.2
python-dotenv~=1.1.0
pandas~=2.2.3
pytesseract~=0.3.13